pyproject.toml
[project]
name = "pdf-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu>=1.10.0",
    "hf-xet>=1.0.3",
    "nltk>=3.9.1",
    "pymupdf>=1.25.5",
    "rank-bm25>=0.2.2",
    "requests>=2.32.3",
    "sentence-transformers>=4.1.0",
    "streamlit>=1.44.1",
    "ragas>=0.1.0",
    "datasets>=2.14.0",
    "llama-index>=0.12.33",
    "llama-index-llms-groq>=0.3.1",
    "rapidfuzz>=3.13.0",
]
